requirements.txt
Django>=4.2.0,<5.0.0
selenium>=4.10.0
pytest>=7.3.1
pytest-django>=4.5.2
webdriver-manager>=3.8.6
opencv-python>=4.7.0
pytest-html>=3.2.0
pillow>=9.5.0